// autogpt_platform/frontend/src/app/api/__generated__/endpoints/folders/folders.ts
export type getV2ListLibraryFoldersResponse200 = {
  data: FolderListResponse;
  status: 200;
};

export type getV2ListLibraryFoldersResponse401 = {
  data: HTTP401NotAuthenticatedErrorResponse;
  status: 401;
};

export type getV2ListLibraryFoldersResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type getV2ListLibraryFoldersResponse500 = {
  data: void;
  status: 500;
};

export type getV2ListLibraryFoldersResponseSuccess =
  getV2ListLibraryFoldersResponse200 & {
    headers: Headers;
  };